mgr/dashboard: Move some e2e into new "ui" folder
