+++ src/pybind/mgr/dashboard/frontend/e2e/ui/user-mgmt.e2e-spec.ts
@@ -1,4 +1,4 @@
-import { Helper } from './helper.po';
+import { Helper } from '../helper.po';
 import { UserMgmtPageHelper } from './user-mgmt.po';
 
 describe('User Management page', () => {
